creado el make

+++ main.c
@@ -5,9 +5,9 @@ int main ()
 {
     char string[] = "Leonel";
     char string2[] = "Lionell";
-    printf("Resultado con mi funcion: %d\n", ft_strlcpy(string, string2, 8));
+    printf("Resultado con mi funcion: %d\n",(int)ft_strlcpy(string, string2, 8));
     printf("destino: %s\n", string);
-    printf("size dest: %d\n", ft_strlen(string));
+    printf("size dest: %d\n", (int)ft_strlen(string));
     printf("hello git");
     return(0);
 }
